+++ utils/db.js
@@ -1,4 +1,4 @@
 import knex from "knex"
-import { development } from "./knexfile.js"
+import { development } from "../config/knex.js"
 
 export const db = knex(development)
